requirements.txt
pytorch==1.8.0
hydra-core==1.1.2
numpy==1.21.5
pandas==1.4.2
dataclasses==0.7
gdown==4.4.0
scikit-learn==1.0.2